Added find and contain functions, although the interface is wonky.
The 'wonkiness' comes from the fact that find can return a size_t. If it can't find the number, find returns SIZE_MAX, but if it actually does find it, but at position SIZE_MAX, it will also return SIZE_MAX. It would be possible to return an ssize_t from find, but that would mean a type mismatch (and therefore, technically, a limiting of the range) of the capacity and elements types. However, because it's very, very unlikely something will every actually contain size_t elements (because, theoretically, that's the largest number of elements an array in C can hold), this is unlikely to be a major problem for most applications.